fix: add legacy-html-assets generator
Move asset folder copying to its own generator that we can call separately from
the legacy-html generator
