+++ src/generators/legacy-html-assets/assets/api.js
